Adding Python version of harmonization_strategies_v2
